requirements.txt
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
oauthlib==3.2.2
packaging==23.0
pycodestyle==2.10.0
pydantic==1.10.4
python-dotenv==0.21.1
replicate==0.4.0
requests==2.28.2
requests-oauthlib==1.3.1
tweepy==4.12.1
typing_extensions==4.4.0
urllib3==1.26.14